Ai-image-Backend/requirements.txt
# Flask Web Framework
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# Database
pymongo==4.6.1

# Environment Configuration
python-dotenv==1.0.0

# HTTP Requests
requests==2.32.3

# AI/ML APIs
anthropic==0.40.0
google-genai

# Image Processing
Pillow==10.1.0

# Azure Blob Storage
azure-storage-blob==12.19.0
